+++ backend/manager/modules/dal/src/main/java/org/ovirt/engine/core/dao/DiskImageDaoImpl.java
@@ -247,9 +247,9 @@ public class DiskImageDaoImpl extends BaseDao implements DiskImageDao {
             entity.setId(getGuidDefaultEmpty(rs, "image_group_id"));
             entity.setStoragePoolId(getGuid(rs, "storage_pool_id"));
             entity.setReadRate(rs.getInt("read_rate"));
-            entity.setReadOps(rs.getInt("read_ops"));
+            entity.setReadOps(rs.getLong("read_ops"));
             entity.setWriteRate(rs.getInt("write_rate"));
-            entity.setWriteOps(rs.getInt("write_ops"));
+            entity.setWriteOps(rs.getLong("write_ops"));
             entity.setImageTransferPhase(rs.getObject("image_transfer_phase") != null
                     ? ImageTransferPhase.forValue(rs.getInt("image_transfer_phase")) : null);
             entity.setTransferType(TransferType.forValue(rs.getInt("image_transfer_type")));
